pyproject.toml
[project]
name = "ai-friend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "edge-tts>=7.2.8",
    "live2d-py>=0.6.1.1",
    "pyaudio>=0.2.14",
    "pyside6>=6.10.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "speechrecognition>=3.15.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
